+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added @phpstan-require-implements constraints
+
 ## v0.11.14 (2024-08-21)
 * Converted consts to protected PascalCase
 * Updated Veneer dependency and Stub