+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+* Upgraded PHPStan to v2
 * Added @phpstan-require-implements constraints
+* Updated dependencies
+* Tidied boolean logic
+* Fixed Exceptional syntax
 * Added PHP8.4 to CI workflow
+* Made PHP8.4 minimum version
 
 ## v0.11.14 (2024-08-21)
 * Converted consts to protected PascalCase